Fix: product is uint64

+++ util/random.go
@@ -10,8 +10,8 @@ type Random struct {
 
 func (r *Random) Next() uint32 {
 	const a = 16807
-	product := r.seed * a
-	r.seed = (product >> 31) + (product & m)
+	product := uint64(r.seed * a)
+	r.seed = uint32((product >> 31) + (product & uint64(m)))
 	if r.seed > m {
 		r.seed -= m
 	}
